# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GSS Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegistrationResult {
  success: boolean;
  message: string;
  id: number;
  accountNumber?: string;
}

export interface CompanyRegistrationBody {
  companyName: string;
  commercialRegister?: string;
  city: string;
  address?: string;
  branchCount?: number;
  activityType?: string;
  contactName: string;
  jobTitle?: string;
  department?: string;
  phone: string;
  email: string;
  authorizationConfirmed?: boolean;
  agreementConfirmed?: boolean;
  requestedServices?: string[];
  currentVendors?: string;
  operationVolume?: string;
  collaborationModel: string;
}

export interface VendorRegistrationBody {
  vendorType: string;
  name: string;
  phone: string;
  email: string;
  services: string[];
  serviceScope: string;
  financialModelAccepted?: boolean;
}

export interface ConsultantRegistrationBody {
  name: string;
  city: string;
  phone: string;
  email: string;
  expertiseArea: string;
  contributionType?: string;
  yearsOfExperience?: number;
  revenueModelAccepted?: boolean;
}

export interface ContactBody {
  name: string;
  phone?: string;
  email: string;
  subject?: string;
  message: string;
}

export interface ServiceRequest {
  id: number;
  title: string;
  status: string;
  serviceType: string;
  createdAt: string;
  updatedAt: string;
}

export interface CompanyDashboard {
  totalRequests: number;
  activeRequests: number;
  completedRequests: number;
  pendingRequests: number;
  recentRequests: ServiceRequest[];
  monthlyCostSavings: number;
  vendorRatingsAvg: number;
}

export interface AssignedRequest {
  id: number;
  title: string;
  companyName: string;
  status: string;
  serviceType: string;
  assignedAt: string;
}

export interface VendorDashboard {
  totalAssigned: number;
  activeJobs: number;
  completedJobs: number;
  pendingQuotes: number;
  assignedRequests: AssignedRequest[];
  performanceScore: number;
}

export interface ConsultationItem {
  id: number;
  title: string;
  status: string;
  linkedCompany: string;
  rewardAmount: number;
  createdAt: string;
}

export interface ConsultantDashboard {
  totalConsultations: number;
  activeConsultations: number;
  completedConsultations: number;
  totalEarnings: number;
  pendingEarnings: number;
  consultations: ConsultationItem[];
}